Create glossary.php
This web page queried a mysql database I created to store technical definitions for a client and displayed it in a browser. Site used jquery , javascript, ajax, and css to make it interactive--clicking on a letter would just display those definitions starting with that letter - clicking on a term would expand/display that definition. Still need to locate supporting files, but this is what I was able to retrieve.